glwidgets: Make the 'Compare with mesh' object static, even while live-transformed during the roto-translate tool operation; we want to keep it unaffected, only moving the collision itself and having this static for reference. Suggested by @Dalion on Discord.
By the time renderBrfItem() was called, it was already too late, as the glMultMatrixf() call had already happened, and it we did a glPopMatrix() inside it we would need to push an identical one afterwards with the same (i==lastSelected) || applyExtraMatrixToAll logic anyway, and we didn't have access to the selected index to see if were the last one. So, instead of duplicating and countering the logic, move the comparison mesh drawing one level up, the function uses a C++ template anyway, so there are multiple copies for each BRF asset type, and this one would only work for BrfBody ones (collisions). Make it simpler.

Dalion
 —
Yesterday at 10:27 PM
@Swyter just found two potential further improvements to openbrf:
1) on collision tab, when you have compare with mesh checkbox ticked and try to rescale it, the mesh is rescaled too, which makes it difficult to adjust the size. When you confirm the changes the mesh size resets anyway so I'm not sure why it would be affected at all, looks like a bug
2) it would be great to introduce a position shift copying from one mesh to another. Like here I have to select both balls and move the centered one with apply to last selected object only to put it on the same spot. Would be faster to select shifted one - edit - copy, select centered one - edit - paste position