askTransformDialog: Reduce the step size of the rotation boxes in the roto-translate tool from +/-15 degrees per click to +/-10 deg for X and Y, and +/-5 for Z, as that one needs more granularity. Requested by @InVain on Discord.
InVain
 —
04/29/2024 8:52 PM
Swyter, you never cease to amaze. I was so used already to closing the menu and checking re-checking the mesh that I had all but forgotten that it's missing this feature 🙂
When I last gave this a try, I noticed that the translate tool works with different iterations that the old version. Could you look into that? I find that the old tool had perfect stepping.
(Maybe same for the UV menu, but I'm not sure, only tried it once.)