main_create, selector: Make a few hardcoded English strings translatable. Suggested by @tos1:
It is not my intention to offend or shame you. I look forward to its progress and further evolution as I learn and test the OpenBRF.

Below is a list of problem I am currently aware of regarding the tr() part. These should give not a little impact on each (active/inactive/potential) translator. (I don't know whether some intended ones are included.)

--------------------
- sceen -> scene [file: askNewUiPictureDialog.ui]
- Shininiess -> Shininess [file: glwidgets.cpp]
- Thinkness -> Thickness (or Thinness)?
- Elnongate -> Elongate [file: guipanel.ui]
- Brighness: -> Brightness: [file: not askHueSatBriDialog.ui but guipanel.ui. Is this used??]
- Enable disable lighting, -> . (at end of line) [file: guipanel.ui]
- Click to open file containing this bumpmap, -> . (at end of line) [file: guipanel.ui]
--------------------
Recent Modules -> Missing tr() [file: main_create.cpp]
--------------------
Rotat: -> (intended for shortening??) [file: guipanel.ui:]
--------------------

Also, I'm concerned that the left/right of so many labels in dialogs/panels are missing. For example, it is understandable that the "frames:" of the mesh cannot fit, but there seems to be enough width to display the "faces:". I think the placeholder should be widen.

label1a.jpg

It may depend on the environment. In my case, I'm seeing an English font made for Windows Japanese version, so I should check for some other cases (font, size, etc.). FYI, I'm using <small></small> for the Japanese of "frames:" (right side of the image above).

Thanks.