guipanel: Support smart multi-selection while in the shader mode with the panel fields reacting dynamically to it, disabling the texture access fields when more than one entry is selected at the same time. Essentially mirroring the behavior in mesh or material mode.
                 Suggested by, you guessed it, @Erundil via the TaleWorlds Forums.

https://forums.taleworlds.com/index.php?threads/openbrf-redux-2025-04-12-%E2%80%94-a-de-rusted-maintained-bug-fixed-version-of-openbrf-with-64-bit-support-and-dark-mode.462517/page-4#post-9914613

 Erundil
Erundil
Sergeant

    Yesterday at 10:16 PM

    Add bookmark
    #70

Yay, I can select multiple shaders now! And all the previously reported shader-related problems are gone!

@Swyter one thing to report about that though.
When I select multiple objects, the Data frame normally shows only data that is identical for all selected objects, fields that differ are emptied (try on meshes, or materials to reproduce).
When selecting multiple shaders however, the Data frame shows data of the 1st selected shader (not necessarily 1st on the list, by 1st I mean the one from which I started the multiselect), without analyzing if it matches the other shaders. It's a big discrepancy with how it works for other objects (meshes, materials, textures etc.) so it will confuse users if left as is.